requirements.txt
Flask==2.0.1
gunicorn==20.1.0
PyMuPDF==1.19.0
Pillow==8.3.1